Resolving merging conflicts
